No need for SIMD
We were only using the 128-bit vector because we needed to pack 6 7-bit integers together. We could have as well made our own struct for this but now there's a 128-bit Int, it suffices just fine.
Use of SIMD would imply some further optimization, but there are no benefits here.